@@ -33,10 +33,10 @@ export function reducer(state = initialState, action: BlogActionsUnion): State {
         ...state,
         selectedBlogId: undefined,
         errorMessage: undefined,
-      }
+      };
     }
     case BlogActionTypes.LOAD_BLOGS_WITH_QUERY_SUCCESS: {
-      const response: HttpResponse<Object> = action.payload;
+      const response: HttpResponse<any> = action.payload;
       const blogs: Blog[] = map(response.body, (blog: any) => new Blog(blog));
       const lastPageLink: string = find(
         split(response.headers.get('Link'), ','),
